More massive changes.   Modifie dthe database so that S_List has a Trim Value type decimal(12,4).
I added the class Trim to aids adding the trim value to the processed value.   There were changes to several classes and methods as well.

Also I believe I might have found the comms problem.   timer2 needed to be of a longer duration when GetHumps() was issued.  I set it tp 1.5 secs during GetHumps() and abck to .1 sec for all the rest. It takes the device alomost a second to respond.